learning-steps/my_sprocs/requirements.txt
snowflake-snowpark-python
pandas
openpyxl
